% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\docType{data}
\name{kg_lookup}
\alias{kg_lookup}
\title{Köppen-Geiger lookup table (internal)}
\format{
An object of class \code{character} of length 30.
}
\usage{
kg_lookup
}
\description{
Köppen-Geiger lookup table (internal)
}
\keyword{datasets}
