% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\docType{data}
\name{lc_lookup}
\alias{lc_lookup}
\title{Land cover lookup table (internal)}
\format{
An object of class \code{character} of length 8.
}
\usage{
lc_lookup
}
\description{
Land cover lookup table (internal)
}
\keyword{datasets}
