% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_impute.R
\name{impute_medians}
\alias{impute_medians}
\title{Replace NA values in selected columns with column medians.}
\usage{
impute_medians(df, numeric_cols)
}
\arguments{
\item{df}{Data frame.}

\item{numeric_cols}{Character vector of column names to impute.}
}
\value{
Data frame with NAs replaced.
}
\description{
Replace NA values in selected columns with column medians.
}
